introduced vectorizer class
