+++ l5kit/l5kit/data/__init__.py
@@ -1,7 +1,7 @@
 from .combine import get_combined_scenes
 from .filter import (filter_agents_by_frames, filter_agents_by_labels, filter_agents_by_track_id,
                      filter_tl_faces_by_frames, filter_tl_faces_by_status, get_agents_slice_from_frames,
-                     get_frames_slice_from_scenes, get_tl_faces_slice_from_frames)
+                     get_frames_slice_from_scenes, get_tl_faces_slice_from_frames, filter_agents_by_distance)
 from .labels import PERCEPTION_LABEL_TO_INDEX, PERCEPTION_LABELS, TL_FACE_LABEL_TO_INDEX, TL_FACE_LABELS
 from .local_data_manager import DataManager, LocalDataManager
 from .map_api import MapAPI
@@ -22,6 +22,7 @@ __all__ = [
     "PERCEPTION_LABEL_TO_INDEX",
     "filter_agents_by_labels",
     "filter_agents_by_frames",
+    "filter_agents_by_distance",
     "filter_tl_faces_by_frames",
     "MapAPI",
     "zarr_concat",
