+++ l5kit/l5kit/data/filter.py
@@ -58,6 +58,21 @@ def filter_agents_by_labels(agents: np.ndarray, threshold: float = 0.5) -> np.nd
     return agents[label_indices]
 
 
+def filter_agents_by_distance(agents: np.ndarray, centroid: np.ndarray, max_distance: float) -> np.ndarray: 
+    """Filter agents by distance, cut to `max_distance` and sort the result
+    Args:
+        agents (np.ndarray): array of agents
+        centroid (np.ndarray): centroid towards which compute distance
+        max_distance (float): max distance to cut off
+    Returns:
+        np.ndarray: agents sorted and cut to max_distance
+    """
+    agents_dist = np.linalg.norm(agents["centroid"] - centroid, axis=-1)
+    agents = agents[agents_dist < max_distance]
+    agents_dist = agents_dist[agents_dist < max_distance]
+    agents = agents[np.argsort(agents_dist)]
+    return agents
+
 def filter_agents_by_track_id(agents: np.ndarray, track_id: int) -> np.ndarray:
     """Return all agent object (np.ndarray) of a given track_id.
 
@@ -169,3 +184,26 @@ def get_tl_faces_slice_from_frames(frame_a: np.ndarray, frame_b: Optional[np.nda
         frame_b = frame_a
     tl_faces_index_end = frame_b["traffic_light_faces_index_interval"][1]
     return slice(tl_faces_index_start, tl_faces_index_end)
+
+# TODO @lberg: we're missing the AV
+def get_other_agents_ids(
+    all_agents_ids: np.ndarray, priority_ids: np.ndarray, selected_track_id: Optional[int], max_agents: int
+) -> List[np.uint64]:
+    """Get ids of agents around selected_track_id. Give precedence to `priority_ids`
+    over `all_agents_ids` and cut to `max_agents`
+    Args:
+        all_agents_ids (np.ndarray): ids of all the agents from present to past
+        priority_ids (np.ndarray): ids of agents we know are reliable in the present
+        selected_track_id (Optional[int]): current id of the agent of interest
+        max_agents (int): max agents to take
+    Returns:
+        List[np.uint64]: the list of track ids of agents to take
+    """
+    agents_taken: Set[np.uint64] = set()
+    # ensure we give priority to reliable, then fill starting from the past
+    for agent_id in np.concatenate([priority_ids, all_agents_ids]):
+        if len(agents_taken) >= max_agents:
+            break
+        if agent_id != selected_track_id:
+            agents_taken.add(agent_id)
+    return list(agents_taken)
