+++ l5kit/l5kit/dataset/ego_vectorized.py
@@ -22,212 +22,7 @@ from l5kit.data.filter import filter_agents_by_track_id
 from l5kit.geometry import compute_agent_pose, rotation33_as_yaw
 from l5kit.rasterization import EGO_EXTENT_HEIGHT, EGO_EXTENT_LENGTH, EGO_EXTENT_WIDTH, RenderContext
 from l5kit.rasterization.semantic_rasterizer import indices_in_bounds
-from l5kit.data.map_api import CACHE_SIZE, InterpolationMethod, MapAPI
-
-def agents_func(history_coords_offset, history_yaws_offset, history_availability, history_agents, agent_centroid_m, max_agents_distance, selected_track_id, other_agents_num, history_num_frames_max, future_num_frames, history_frames, agent_from_world, agent_yaw_rad, future_frames, future_agents, history_num_frames_agents):
-    # compute agent features
-    # sequence_length x 2 (two being x, y)
-    agent_points = history_coords_offset.copy()
-    # sequence_length x 1
-    agent_yaws = history_yaws_offset.copy()
-    # sequence_length x xy+yaw (3)
-    agent_trajectory_polyline = np.concatenate([agent_points, agent_yaws], axis=-1)
-    agent_polyline_availability = history_availability.copy()
-
-    # get agents around AoI sorted by distance in a given radius. Give priority to agents in the current time step
-    history_agents_flat = filter_agents_by_labels(np.concatenate(history_agents))
-    history_agents_flat = filter_agents_by_distance(history_agents_flat, agent_centroid_m, max_agents_distance)
-
-    cur_agents = filter_agents_by_labels(history_agents[0])
-    cur_agents = filter_agents_by_distance(cur_agents, agent_centroid_m, max_agents_distance)
-
-    list_agents_to_take = get_other_agents_ids(
-        history_agents_flat["track_id"], cur_agents["track_id"], selected_track_id, other_agents_num
-    )
-
-    # Loop to grab history and future for all other agents
-    all_other_agents_history_positions = np.zeros((other_agents_num, history_num_frames_max + 1, 2), dtype=np.float32)
-    all_other_agents_history_yaws = np.zeros((other_agents_num, history_num_frames_max + 1, 1), dtype=np.float32)
-    all_other_agents_history_extents = np.zeros((other_agents_num, history_num_frames_max + 1, 2), dtype=np.float32)
-    all_other_agents_history_availability = np.zeros((other_agents_num, history_num_frames_max + 1), dtype=np.float32)
-    all_other_agents_types = np.zeros((other_agents_num,), dtype=np.int64)
-
-    all_other_agents_future_positions = np.zeros((other_agents_num, future_num_frames, 2), dtype=np.float32)
-    all_other_agents_future_yaws = np.zeros((other_agents_num, future_num_frames, 1), dtype=np.float32)
-    all_other_agents_future_extents = np.zeros((other_agents_num, future_num_frames, 2), dtype=np.float32)
-    all_other_agents_future_availability = np.zeros((other_agents_num, future_num_frames), dtype=np.float32)
-
-    for idx, track_id in enumerate(list_agents_to_take):
-        (
-            agent_history_coords_offset,
-            agent_history_yaws_offset,
-            agent_history_extent,
-            agent_history_availability,
-        ) = get_relative_poses(
-            history_num_frames_max + 1, history_frames, track_id, history_agents, agent_from_world, agent_yaw_rad
-        )
-
-        all_other_agents_history_positions[idx] = agent_history_coords_offset
-        all_other_agents_history_yaws[idx] = agent_history_yaws_offset
-        all_other_agents_history_extents[idx] = agent_history_extent
-        all_other_agents_history_availability[idx] = agent_history_availability
-        # NOTE (@lberg): assumption is that an agent doesn't change class (seems reasonable)
-        # We look from history backward and choose the most recent time the track_id was available.
-        current_other_actor = filter_agents_by_track_id(history_agents_flat, track_id)[0]
-        all_other_agents_types[idx] = np.argmax(current_other_actor["label_probabilities"])
-
-        (
-            agent_future_coords_offset,
-            agent_future_yaws_offset,
-            agent_future_extent,
-            agent_future_availability,
-        ) = get_relative_poses(
-            future_num_frames, future_frames, track_id, future_agents, agent_from_world, agent_yaw_rad
-        )
-        all_other_agents_future_positions[idx] = agent_future_coords_offset
-        all_other_agents_future_yaws[idx] = agent_future_yaws_offset
-        all_other_agents_future_extents[idx] = agent_future_extent
-        all_other_agents_future_availability[idx] = agent_future_availability
-
-    # crop similar to ego above
-    all_other_agents_history_positions[:, history_num_frames_agents + 1 :] *= 0
-    all_other_agents_history_yaws[:, history_num_frames_agents + 1 :] *= 0
-    all_other_agents_history_extents[:, history_num_frames_agents + 1 :] *= 0
-    all_other_agents_history_availability[:, history_num_frames_agents + 1 :] *= 0
-
-    # compute other agents features
-    # num_other_agents (M) x sequence_length x 2 (two being x, y)
-    agents_points = all_other_agents_history_positions.copy()
-    # num_other_agents (M) x sequence_length x 1
-    agents_yaws = all_other_agents_history_yaws.copy()
-    # agents_extents = all_other_agents_history_extents[:, :-1]
-    # num_other_agents (M) x sequence_length x self._vector_length
-    other_agents_polyline = np.concatenate([agents_points, agents_yaws], axis=-1)
-    other_agents_polyline_availability = all_other_agents_history_availability.copy()
-
-    agent_dict = {
-        "all_other_agents_history_positions": all_other_agents_history_positions,
-        "all_other_agents_history_yaws": all_other_agents_history_yaws,
-        "all_other_agents_history_extents": all_other_agents_history_extents,
-        "all_other_agents_history_availability": all_other_agents_history_availability.astype(np.bool),
-        "all_other_agents_future_positions": all_other_agents_future_positions,
-        "all_other_agents_future_yaws": all_other_agents_future_yaws,
-        "all_other_agents_future_extents": all_other_agents_future_extents,
-        "all_other_agents_future_availability": all_other_agents_future_availability.astype(np.bool),
-        "all_other_agents_types": all_other_agents_types,
-        "agent_trajectory_polyline": agent_trajectory_polyline,
-        "agent_polyline_availability": agent_polyline_availability.astype(np.bool),
-        "other_agents_polyline": other_agents_polyline,
-        "other_agents_polyline_availability": other_agents_polyline_availability.astype(np.bool),
-    }
-
-    return agent_dict
-
-def map_func(lane_cfg_params, mapAPI, agent_centroid_m, agent_from_world, history_tl_faces):
-      # START WORKING ON LANES
-    # TODO (lberg): this implementation is super ugly, I'll fix it
-    # TODO (anasrferreira): clean up more and add interpolation params to configuration as well.
-    MAX_LANES = lane_cfg_params["max_num_lanes"]
-    MAX_POINTS_LANES = lane_cfg_params["max_points_per_lane"]
-    MAX_POINTS_CW = lane_cfg_params["max_points_per_crosswalk"]
-
-    MAX_LANE_DISTANCE = lane_cfg_params["max_retrieval_distance_m"]
-    INTERP_METHOD = InterpolationMethod.INTER_ENSURE_LEN  # split lane polyline by fixed number of points
-    STEP_INTERPOLATION = MAX_POINTS_LANES  # number of points along lane
-    MAX_CROSSWALKS = lane_cfg_params["max_num_crosswalks"]
-
-    lanes_points = np.zeros((MAX_LANES * 2, MAX_POINTS_LANES, 2), dtype=np.float32)
-    lanes_availabilities = np.zeros((MAX_LANES * 2, MAX_POINTS_LANES), dtype=np.float32)
-
-    lanes_mid_points = np.zeros((MAX_LANES, MAX_POINTS_LANES, 2), dtype=np.float32)
-    lanes_mid_availabilities = np.zeros((MAX_LANES, MAX_POINTS_LANES), dtype=np.float32)
-    lanes_tl_feature = np.zeros((MAX_LANES, MAX_POINTS_LANES, 1), dtype=np.float32)
-
-    # 8505 x 2 x 2
-    lanes_bounds = mapAPI.bounds_info["lanes"]["bounds"]
-
-    # filter first by bounds and then by distance, so that we always take the closest lanes
-    lanes_indices = indices_in_bounds(agent_centroid_m, lanes_bounds, MAX_LANE_DISTANCE)
-    distances = []
-    for lane_idx in lanes_indices:
-        lane_id = mapAPI.bounds_info["lanes"]["ids"][lane_idx]
-        lane = mapAPI.get_lane_as_interpolation(lane_id, STEP_INTERPOLATION, INTERP_METHOD)
-        lane_dist = np.linalg.norm(lane["xyz_midlane"][:, :2] - agent_centroid_m, axis=-1)
-        distances.append(np.min(lane_dist))
-    lanes_indices = lanes_indices[np.argsort(distances)]
-
-     # TODO: move below after traffic lights
-    crosswalks_bounds = mapAPI.bounds_info["crosswalks"]["bounds"]
-    crosswalks_indices = indices_in_bounds(agent_centroid_m, crosswalks_bounds, MAX_LANE_DISTANCE)
-    crosswalks_points = np.zeros((MAX_CROSSWALKS, MAX_POINTS_CW, 2), dtype=np.float32)
-    crosswalks_availabilities = np.zeros_like(crosswalks_points[..., 0])
-    for i, xw_idx in enumerate(crosswalks_indices[:MAX_CROSSWALKS]):
-        xw_id = mapAPI.bounds_info["crosswalks"]["ids"][xw_idx]
-        points = mapAPI.get_crosswalk_coords(xw_id)["xyz"]
-        points = transform_points(points[:MAX_POINTS_CW, :2], agent_from_world)
-        n = len(points)
-        crosswalks_points[i, :n] = points
-        crosswalks_availabilities[i, :n] = True
-
-    active_tl_faces = set(filter_tl_faces_by_status(history_tl_faces[0], "ACTIVE")["face_id"].tolist())
-    active_tl_face_to_color: Dict[str, str] = {}
-    for face in active_tl_faces:
-        try:
-            active_tl_face_to_color[face] = get_color_for_face(mapAPI, face)
-        except KeyError:
-            continue  # this happens only on KIRBY, 2 TLs have no match in the map
-
-    for out_idx, lane_idx in enumerate(lanes_indices[:MAX_LANES]):
-        lane_id = mapAPI.bounds_info["lanes"]["ids"][lane_idx]
-        lane = mapAPI.get_lane_as_interpolation(lane_id, STEP_INTERPOLATION, INTERP_METHOD)
-
-        xy_left = lane["xyz_left"][:MAX_POINTS_LANES, :2]
-        xy_right = lane["xyz_right"][:MAX_POINTS_LANES, :2]
-        # convert coordinates into local space
-        xy_left = transform_points(xy_left, agent_from_world)
-        xy_right = transform_points(xy_right, agent_from_world)
-
-        num_vectors_left = len(xy_left)
-        num_vectors_right = len(xy_right)
-
-        lanes_points[out_idx * 2, :num_vectors_left] = xy_left
-        lanes_points[out_idx * 2 + 1, :num_vectors_right] = xy_right
-
-        lanes_availabilities[out_idx * 2, :num_vectors_left] = 1
-        lanes_availabilities[out_idx * 2 + 1, :num_vectors_right] = 1
-
-        midlane = lane["xyz_midlane"][:MAX_POINTS_LANES, :2]
-        midlane = transform_points(midlane, agent_from_world)
-        num_vectors_mid = len(midlane)
-
-        lanes_mid_points[out_idx, :num_vectors_mid] = midlane
-        lanes_mid_availabilities[out_idx, :num_vectors_mid] = 1
-
-        lanes_tl_feature[out_idx, :num_vectors_mid] = get_tl_feature_for_lane(mapAPI, lane_id, active_tl_face_to_color)
-
-    # disable all points over the distance threshold
-    valid_distances = np.linalg.norm(lanes_points, axis=-1) < MAX_LANE_DISTANCE
-    lanes_availabilities *= valid_distances
-    valid_mid_distances = np.linalg.norm(lanes_mid_points, axis=-1) < MAX_LANE_DISTANCE
-    lanes_mid_availabilities *= valid_mid_distances
-
-    # 2 MAX_LANES x MAX_VECTORS x (XY + TL-feature)
-    # -> 2 MAX_LANES for left and right
-    lanes = np.concatenate([lanes_points, np.zeros_like(lanes_points[..., [0]])], axis=-1)
-    # pad such that length is 3
-    crosswalks = np.concatenate([crosswalks_points, np.zeros_like(crosswalks_points[..., [0]])], axis=-1)
-    # MAX_LANES x MAX_VECTORS x 3 (XY + 1 TL-feature)
-    lanes_mid = np.concatenate([lanes_mid_points, lanes_tl_feature], axis=-1)
-
-    return {
-        "lanes": lanes,
-        "lanes_availabilities": lanes_availabilities.astype(np.bool),
-        "lanes_mid": lanes_mid,
-        "lanes_mid_availabilities": lanes_mid_availabilities.astype(np.bool),
-        "crosswalks": crosswalks,
-        "crosswalks_availabilities": crosswalks_availabilities.astype(np.bool),
-    }
-
+from l5kit.vectorization.vectorizer import Vectorizer
 
 
 def vectorized_generate_agent_sample(
@@ -242,10 +37,7 @@ def vectorized_generate_agent_sample(
     future_num_frames: int,
     step_time: float,
     filter_agents_threshold: float,
-    other_agents_num: int,
-    max_agents_distance: float,
-    mapAPI: MapAPI,
-    lane_cfg_params: dict,
+    vectorizer: Vectorizer,
     rasterizer: Optional[Rasterizer] = None,
     perturbation: Optional[Perturbation] = None,
 ) -> dict:
@@ -330,7 +122,7 @@ def vectorized_generate_agent_sample(
         "raster_from_world": raster_from_world,
         "agent_from_world": agent_from_world,
         "world_from_agent": world_from_agent,
-            "target_positions": future_coords_offset,
+        "target_positions": future_coords_offset,
         "target_yaws": future_yaws_offset,
         "target_extents": future_extents,
         "target_availabilities": future_availability.astype(np.bool),
@@ -343,79 +135,9 @@ def vectorized_generate_agent_sample(
         "speed": np.linalg.norm(future_vels_mps[0]),
     }
 
-    agent_dict = agents_func(history_coords_offset, history_yaws_offset, history_availability, history_agents, agent_centroid_m, max_agents_distance, selected_track_id, other_agents_num, history_num_frames_max, future_num_frames, history_frames, agent_from_world, agent_yaw_rad, future_frames, future_agents, history_num_frames_agents)
-    map_dict = map_func(lane_cfg_params, mapAPI, agent_centroid_m, agent_from_world, history_tl_faces)
+    vec_dict = vectorizer.vectorize(selected_track_id, agent_centroid_m, agent_yaw_rad, agent_from_world, history_frames, history_agents, history_tl_faces, history_coords_offset, history_yaws_offset, history_availability, future_frames, future_agents)
 
-    return {**d1, **agent_dict, **map_dict}
-
-def filter_agents_by_distance(agents: np.ndarray, centroid: np.ndarray, max_distance: float) -> np.ndarray:
-    """Filter agents by distance, cut to `max_distance` and sort the result
-    Args:
-        agents (np.ndarray): array of agents
-        centroid (np.ndarray): centroid towards which compute distance
-        max_distance (float): max distance to cut off
-    Returns:
-        np.ndarray: agents sorted and cut to max_distance
-    """
-    agents_dist = np.linalg.norm(agents["centroid"] - centroid, axis=-1)
-    agents = agents[agents_dist < max_distance]
-    agents_dist = agents_dist[agents_dist < max_distance]
-    agents = agents[np.argsort(agents_dist)]
-    return agents
-
-def get_tl_feature_for_lane(map_api, lane_id, active_tl_face_to_color) -> int:
-    """ Get traffic light feature for a lane given its active tl faces and a constant priority map.
-    """
-    # Map from traffic light state to its feature and priority index (to disambiguate multiple active tl faces)
-    # "None" state is special and mean that a lane does not have a traffic light. "Unknown" means that the traffic
-    # light exists but PCP cannot detect its state.
-    # Except for `none`, priority increases with numbers
-    tl_color_to_priority_idx = {"unknown": 0, "green": 1, "yellow": 2, "red": 3, "none": 4}
-
-    lane_tces = map_api.get_lane_traffic_control_ids(lane_id)
-    lane_tls = [tce for tce in lane_tces if map_api.is_traffic_light(tce)]
-    if len(lane_tls) == 0:
-        return tl_color_to_priority_idx["none"]
-
-    active_tl_faces = active_tl_face_to_color.keys() & lane_tces
-    # The active color with higher priority is kept
-    highest_priority_idx_active = tl_color_to_priority_idx["unknown"]
-    for active_tl_face in active_tl_faces:
-        active_color = active_tl_face_to_color[active_tl_face]
-        highest_priority_idx_active = max(highest_priority_idx_active, tl_color_to_priority_idx[active_color])
-    return highest_priority_idx_active
-    
-@lru_cache(maxsize=CACHE_SIZE)
-def get_color_for_face(mapAPI: MapAPI, face_id: str) -> str:
-    """Obtain the color for the given face_id"""
-    for color in ["red", "yellow", "green"]:
-        if mapAPI.is_traffic_face_color(face_id, color):
-            return color
-    assert False, f"Face {face_id} has no valid color."
-    return 0
-
-# TODO @lberg: we're missing the AV
-def get_other_agents_ids(
-    all_agents_ids: np.ndarray, priority_ids: np.ndarray, selected_track_id: Optional[int], max_agents: int
-) -> List[np.uint64]:
-    """Get ids of agents around selected_track_id. Give precedence to `priority_ids`
-    over `all_agents_ids` and cut to `max_agents`
-    Args:
-        all_agents_ids (np.ndarray): ids of all the agents from present to past
-        priority_ids (np.ndarray): ids of agents we know are reliable in the present
-        selected_track_id (Optional[int]): current id of the agent of interest
-        max_agents (int): max agents to take
-    Returns:
-        List[np.uint64]: the list of track ids of agents to take
-    """
-    agents_taken: Set[np.uint64] = set()
-    # ensure we give priority to reliable, then fill starting from the past
-    for agent_id in np.concatenate([priority_ids, all_agents_ids]):
-        if len(agents_taken) >= max_agents:
-            break
-        if agent_id != selected_track_id:
-            agents_taken.add(agent_id)
-    return list(agents_taken)
+    return {**d1, **vec_dict}
 
 
 class EgoDatasetVectorized(EgoDataset):
@@ -424,11 +146,10 @@ class EgoDatasetVectorized(EgoDataset):
         cfg: dict,
         zarr_dataset: ChunkedDataset,
         rasterizer: Rasterizer,
-        mapAPI: MapAPI,
+        vectorizer: Vectorizer,
         perturbation: Optional[Perturbation] = None,
     ):
         super().__init__(cfg, zarr_dataset, rasterizer, perturbation)
-        self.mapAPI = mapAPI
 
         # replace the sample function to access other agents
         render_context = RenderContext(
@@ -445,14 +166,11 @@ class EgoDatasetVectorized(EgoDataset):
             future_num_frames=cfg["model_params"]["future_num_frames"],
             step_time=cfg["model_params"]["step_time"],
             filter_agents_threshold=cfg["raster_params"]["filter_agents_threshold"],
-            other_agents_num=cfg["data_generation_params"]["other_agents_num"],
-            max_agents_distance=cfg["data_generation_params"]["max_agents_distance"],
-            mapAPI=mapAPI,
-            lane_cfg_params=cfg["data_generation_params"]["lane_params"],
             rasterizer=rasterizer,
             perturbation=perturbation,
+            vectorizer=vectorizer
         )
 
     def get_scene_dataset(self, scene_index: int) -> "EgoDatasetVectorized":
         dataset = super().get_scene_dataset(scene_index).dataset
-        return EgoDatasetVectorized(self.cfg, dataset, self.rasterizer, self.mapAPI, self.perturbation)
+        return EgoDatasetVectorized(self.cfg, dataset, self.rasterizer, self.perturbation)
